Fix: Response parsing fix
parseResponse fails while reading the data from response body if the ctx
is canceled. We do actually read the proper data that we need but still
throw error in the Body is closed.

This PR fixes that (in a way), by using `io.ReadAll` to read the body
and the create a buf from the bytes read.

Keeps existing functionality intact.

Signed-off-by: jay-dee7 <jasdeepsingh.uppal@gmail.com>